+++ src/core/modules/clicks/clicks.js
@@ -6,17 +6,17 @@ import ViewClass from '../../components/view/view-class';
 
 function initClicks(app) {
   function handleClicks(e) {
-    const clicked = $(e.target);
-    const clickedLink = clicked.closest('a');
-    const isLink = clickedLink.length > 0;
-    const url = isLink && clickedLink.attr('href');
-    const isTabLink = isLink && clickedLink.hasClass('tab-link') && (clickedLink.attr('data-tab') || (url && url.indexOf('#') === 0));
+    const $clickedEl = $(e.target);
+    const $clickedLinkEl = $clickedEl.closest('a');
+    const isLink = $clickedLinkEl.length > 0;
+    const url = isLink && $clickedLinkEl.attr('href');
+    const isTabLink = isLink && $clickedLinkEl.hasClass('tab-link') && ($clickedLinkEl.attr('data-tab') || (url && url.indexOf('#') === 0));
 
     // Check if link is external
     if (isLink) {
       // eslint-disable-next-line
-      if (clickedLink.is(app.params.clicks.externalLinks) || (url && url.indexOf('javascript:') >= 0)) {
-        const target = clickedLink.attr('target');
+      if ($clickedLinkEl.is(app.params.clicks.externalLinks) || (url && url.indexOf('javascript:') >= 0)) {
+        const target = $clickedLinkEl.attr('target');
         if (
           url
           && window.cordova
@@ -35,7 +35,7 @@ function initClicks(app) {
       const moduleClicks = app.modules[moduleName].clicks;
       if (!moduleClicks) return;
       Object.keys(moduleClicks).forEach((clickSelector) => {
-        const matchingClickedElement = clicked.closest(clickSelector).eq(0);
+        const matchingClickedElement = $clickedEl.closest(clickSelector).eq(0);
         if (matchingClickedElement.length > 0) {
           moduleClicks[clickSelector].call(app, matchingClickedElement, matchingClickedElement.dataset());
         }
@@ -46,16 +46,16 @@ function initClicks(app) {
     let clickedLinkData = {};
     if (isLink) {
       e.preventDefault();
-      clickedLinkData = clickedLink.dataset();
+      clickedLinkData = $clickedLinkEl.dataset();
     }
     const validUrl = url && url.length > 0 && url !== '#' && !isTabLink;
-    if (validUrl || clickedLink.hasClass('back')) {
+    if (validUrl || $clickedLinkEl.hasClass('back')) {
       let view;
       if (clickedLinkData.view) {
         view = $(clickedLinkData.view)[0].f7View;
       } else {
-        view = clicked.parents('.view')[0] && clicked.parents('.view')[0].f7View;
-        if (!clickedLink.hasClass('back') && view && view.params.linksView) {
+        view = $clickedEl.parents('.view')[0] && $clickedEl.parents('.view')[0].f7View;
+        if (!$clickedLinkEl.hasClass('back') && view && view.params.linksView) {
           if (typeof view.params.linksView === 'string') view = $(view.params.linksView)[0].f7View;
           else if (view.params.linksView instanceof ViewClass) view = view.params.linksView;
         }
@@ -71,7 +71,10 @@ function initClicks(app) {
           // something wrong there
         }
       }
-      if (clickedLink.hasClass('back')) view.router.back(url, clickedLinkData);
+      if ($clickedLinkEl[0].f7RouteProps) {
+        clickedLinkData.props = $clickedLinkEl[0].f7RouteProps;
+      }
+      if ($clickedLinkEl.hasClass('back')) view.router.back(url, clickedLinkData);
       else view.router.navigate(url, clickedLinkData);
     }
   }