+++ src/phenome/components/list-button.jsx
@@ -88,10 +88,27 @@ export default {
     Utils.bindMethods(this, ['onClick'])
   },
   componentDidMount() {
-    this.refs.linkEl.addEventListener('click', this.onClick);
+    const self = this;
+    const linkEl = self.refs.linkEl;
+    const { routeProps } = self.props;
+    if (routeProps) {
+      linkEl.f7RouteProps = routeProps;
+    }
+    linkEl.addEventListener('click', self.onClick);
+  },
+  componentDidUpdate() {
+    const self = this;
+    const linkEl = self.refs.linkEl;
+    const { routeProps } = self.props;
+    if (routeProps) {
+      linkEl.f7RouteProps = routeProps;
+    }
   },
   componentWillUnmount() {
-    this.refs.linkEl.removeEventListener('click', this.onClick);
+    const self = this;
+    const linkEl = self.refs.linkEl;
+    linkEl.removeEventListener('click', this.onClick);
+    delete linkEl.f7RouteProps;
   },
   methods: {
     onClick(event) {